# packages/backend/src/openapi/spec.yaml
openapi: 3.0.3
info:
  title: Jyotish API
  description: |
    Vedic Astrology API for birth chart calculation and AI-powered readings.

    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <token>
    ```

    ## Rate Limiting
    API requests are rate limited. Check response headers for limit information.
  version: 1.0.0
  contact:
    name: Jyotish Support
  license:
    name: MIT
servers:
  - url: /
    description: Current server (relative URL)
  - url: http://localhost:3001
    description: Local development
  - url: http://192.168.1.225:3001
    description: LAN development
...
paths:
...
  /api/subscription/status:
    get:
      summary: Get subscription status
      operationId: getSubscriptionStatus
      tags:
        - Subscription
      parameters:
        - name: userId
          in: query
          required: false
          schema:
            type: string
          description: User ID (optional for authenticated requests)
      responses:
        '200':
          description: Subscription status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionStatusResponse'
  /api/subscription/webhook:
    post:
      summary: RevenueCat webhook endpoint
      operationId: handleSubscriptionWebhook
      tags:
        - Subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RevenueCatWebhookPayload'
      responses:
        '200':
          description: Webhook received
          content:
            application/json:
              schema:
                type: object
                properties:
                  received:
                    type: boolean
                  message:
                    type: string
        '400':
          description: Invalid webhook signature
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/subscription/question-usage:
    get:
      summary: Get question usage for a user
      operationId: getQuestionUsage
      tags:
        - Subscription
      parameters:
        - name: userId
          in: query
          required: false
          schema:
            type: string
          description: User ID
      responses:
        '200':
          description: Question usage info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionUsageResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SubscriptionStatusResponse:
      type: object
      properties:
        success:
          type: boolean
        subscription:
          type: object
          properties:
            isPremium:
              type: boolean
            tier:
              type: string
              enum: [free, premium, pro]
            expiresAt:
              type: string
              format: date-time
              nullable: true
            provider:
              type: string
              nullable: true
              description: Subscription provider (revenuecat, stripe, manual) or null if free
            entitlements:
              type: array
              items:
                type: string
              description: List of entitlements (premium_access, unlimited_questions, all_readings)
    RevenueCatWebhookPayload:
      type: object
      properties:
        api_version:
          type: string
        event:
          type: object
          properties:
            type:
              type: string
              enum: [INITIAL_PURCHASE, RENEWAL, CANCELLATION, UNCANCELLATION, NON_RENEWING_PURCHASE, SUBSCRIPTION_PAUSED, EXPIRATION, BILLING_ISSUE, PRODUCT_CHANGE, TRANSFER]
            app_user_id:
              type: string
            original_app_user_id:
              type: string
            product_id:
              type: string
            entitlement_ids:
              type: array
              items:
                type: string
            expiration_at_ms:
              type: integer
            purchased_at_ms:
              type: integer
            environment:
              type: string
              enum: [SANDBOX, PRODUCTION]
            store:
              type: string
              enum: [APP_STORE, PLAY_STORE, STRIPE, PROMOTIONAL]
    QuestionUsageResponse:
      type: object
      properties:
        success:
          type: boolean
        usage:
          type: object
          properties:
            count:
              type: integer
            remaining:
              type: integer
            limit:
              type: integer
            resetAt:
              type: string
              format: date-time
            isPremium:
              type: boolean

    # Error Schema
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        message:
          type: string
        path:
          type: string